requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0

# Image Processing & Visualization
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Data Augmentation
albumentations>=1.3.0

# HuggingFace Hub
huggingface_hub>=0.23.0

# Progress Bars
tqdm>=4.65.0

# Gradio (for HF Space deployment)
gradio>=4.44.0
